requirements.txt
fpdf==1.7.2
pandas==1.3.3
PyYAML==5.4.1
click==8.0.3